message_fetch: Avoid non-contiguous conversations in recent view.
Fixes #27208, #27207

We only process messages from message list and special data fetched
for recent view only. This avoids us having a conversation
present in recent view which is not contiguous due to loading
a random old conversation.

Also, to ensure displayed conversations are contiguous while
we are loading data, we don't show message list data until we
have found the newest message in the list.