message_notification: Remove unnecessary content truncation.
According to https://shorturl.at/kAHIN, content truncation starts
at 65 characters in some devices and is at max 103 in mac, so 150
was too a high a limit.

But, content is already trimmed and ellipsis is added automatically
by the OS, if it exceeds the limit, so it doesn't seem worth it
to truncate the content ourselves.